+++ src/MyGame/global.cpp
@@ -1,5 +1,5 @@
 #include "global.h"
 
-extern int GLOB_UserAccount = 0;
+extern qint64 GLOB_UserAccount = 0;
 extern QString GLOB_UserName = "";
 extern bool GLOB_IsConnectedToServer = false;
